feat: extract offlineFirst pure data layer, add metro config tests
- #991: Extract getOfflineFirst() as a pure data layer with no React dependencies,
  rewrite useOfflineData hook as a thin wrapper over it
- #990: Add tests for metro.config.js custom resolver and route size analysis
